# Architecture_DZ_10/API/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cloud service for robot-vacuum cleaner",
    "description": "This is a study project considering server based on the OpenAPI 3.0 specification. \nYou can now help us improve the API whether it's by making changes to the definition itself or to the code.\nThat way, with time, we can improve the API in general, and expose some of the new features in OAS3.",
    "contact": {
      "email": "kardan97@yandex.ru"
    },
    "version": "0.0.2"
  },
  "servers": [
    {
      "url": "https://petstore3.swagger.io/api/v3"
    },
    {
      "url": "https://localhost:8080/api/v1"
    }
  ],
...
  "paths": {
    "/robot/add": {
      "post": {
        "tags": [
          "robot"
        ],
        "summary": "Add a new robot to user account",
        "description": "Add a new robot to user account",
        "operationId": "addRobot",
        "requestBody": {
          "description": "Add a new robot to user account",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robot"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Robot"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Robot"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              }
            }
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/robot/{robotId}": {
      "get": {
        "tags": [
          "robot"
        ],
        "summary": "Get robot by ID",
        "description": "Get robot by ID",
        "operationId": "getRobot",
        "parameters": [
          {
            "name": "robotId",
            "in": "path",
            "description": "ID of robot to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Robot not found"
          }
        }
      },
      "put": {
        "tags": [
          "robot"
        ],
        "summary": "Update info about an existing robot",
        "description": "Update info about an existing robot by Id",
        "operationId": "updateRobot",
        "parameters": [
          {
            "name": "robotId",
            "in": "path",
            "description": "ID of robot to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Update an existent robot in the store",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Robot"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Robot"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Robot"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Robot not found"
          }
        }
      },
      "delete": {
        "tags": [
          "robot"
        ],
        "summary": "Delete a robot by ID",
        "description": "delete a robot by ID",
        "operationId": "deleteRobot",
        "parameters": [
          {
            "name": "robotId",
            "in": "path",
            "description": "Robot id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              }
            }
          },
          "400": {
            "description": "Invalid robot ID value"
          },
          "404": {
            "description": "Robot not found"
          }
        }
      }
    },
    "/robot/{robotId}/checkUpdate": {
      "get": {
        "tags": [
          "robot"
        ],
        "summary": "Check a new firmware update for Robot",
        "description": "Check a new firmware update for Robot",
        "operationId": "checkUpdate",
        "parameters": [
          {
            "name": "robotId",
            "in": "path",
            "description": "Robot id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Firmware"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Firmware"
                }
              }
            }
          },
          "400": {
            "description": "Invalid robot ID value"
          },
          "404": {
            "description": "Robot not found"
          }
        }
      }
    },
    "/robot/{robotId}/checkStatus": {
      "get": {
        "tags": [
          "robot"
        ],
        "summary": "Check a current Robot status",
        "description": "Check a current Robot status",
        "operationId": "checkStatus",
        "parameters": [
          {
            "name": "robotId",
            "in": "path",
            "description": "Robot id to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Robot"
                }
              }
            }
          },
          "400": {
            "description": "Invalid robot ID value"
          },
          "404": {
            "description": "Robot not found"
          }
        }
      }
    },
    "/robot/{robotId}/downloadUpdate": {
      "get": {
        "tags": [
          "robot"
        ],
        "summary": "Downloads a new firmware update",
        "description": "Downloads a new firmware update",
        "operationId": "downloadFile",
        "parameters": [
          {
            "name": "robotId",
            "in": "path",
            "description": "ID of pet to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid robot ID value"
          },
          "404": {
            "description": "Robot not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Statistic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "timetableID": {
            "$ref": "#/components/schemas/Timetable"
          },
          "resultID": {
            "type": "string"
          }
        },
        "xml": {
          "name": "statistic"
        }
      },
      "Timetable": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "datetime": {
            "type": "string",
            "format": "date-time"
          },
          "roomID": {
            "$ref": "#/components/schemas/Room"
          },
          "resultID": {
            "$ref": "#/components/schemas/Result"
          }
        },
        "xml": {
          "name": "timetable"
        }
      },
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "roomname": {
            "type": "string"
          },
          "roomsquare": {
            "type": "string"
          }
        },
        "xml": {
          "name": "room"
        }
      },
      "Result": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "Details": {
            "type": "string",
            "description": "What does this code mean"
          }
        },
        "xml": {
          "name": "result"
        }
      },
      "Model": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "firmwareID": {
            "$ref": "#/components/schemas/Firmware"
          }
        },
        "xml": {
          "name": "model"
        }
      },
      "Firmware": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "string"
          },
          "URL": {
            "type": "string"
          }
        },
        "xml": {
          "name": "firmware"
        }
      },
      "Robot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "serialNo": {
            "type": "string"
          },
          "model": {
            "$ref": "#/components/schemas/Model"
          },
          "firmware": {
            "type": "string"
          },
          "config": {
            "type": "array",
            "xml": {
              "wrapped": true
            },
            "items": {
              "type": "string",
              "xml": {
                "name": "config"
              }
            }
          },
          "diadnostic": {
            "type": "array",
            "xml": {
              "wrapped": true
            },
            "items": {
              "type": "integer",
              "format": "int32",
              "xml": {
                "name": "diadnostic"
              }
            }
          },
          "historyClean": {
            "$ref": "#/components/schemas/Statistic"
          },
          "status": {
            "type": "string",
            "description": "robot status in the store",
            "enum": [
              "ready",
              "charging",
              "working"
            ]
          }
        },
        "xml": {
          "name": "robot"
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "xml": {
          "name": "##default"
        }
      }
    },
...
  }
}